# build/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"TaskStatus": {
				"enum": [
					"todo",
					"inprogress",
					"completed"
				],
				"type": "string"
			},
			"Task": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/TaskStatus"
					}
				},
				"required": [
					"id",
					"name",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_ITask.name_": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"CreateTaskDto": {
				"$ref": "#/components/schemas/Pick_ITask.name_"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "tsoa-first-steps",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/tasks": {
			"get": {
				"operationId": "GetTasks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Task"
									},
									"type": "array"
								},
								"examples": {
									"Example 1": {
										"value": [
											{
												"id": "0",
												"name": "An example task",
												"status": "todo"
											}
										]
									}
								}
							}
						}
					}
				},
				"description": "Gets list of all tasks.",
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateTask",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Task"
								},
								"examples": {
									"Example 1": {
										"value": "Successfully added task."
									}
								}
							}
						}
					}
				},
				"description": "Add a new task. Remember that the demo API will not persist this data.",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTaskDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteAllTasks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/tasks/{taskId}": {
			"delete": {
				"operationId": "DeleteTask",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Deletes a task given the id.",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "taskId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"get": {
				"operationId": "GetTaskById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Task"
								}
							}
						}
					}
				},
				"description": "Get the task with the given id.",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "taskId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/tasks/toggle/{taskId}": {
			"put": {
				"operationId": "ToggleTask",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"description": "Toggles the completion status of the task with the given id.",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "taskId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/tasks/markAll/{setting}": {
			"put": {
				"operationId": "ToggleAll",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Toggles the completion status of all tasks to the given status.",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "setting",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/TaskStatus"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}